refactor of dialogs
